+++ include/gen.hpp
@@ -5,7 +5,7 @@
 // n: 長さ
 // chars: 文字種
 template<typename Engine>
-string gen_string(RngT<Engine>& rng, i64 n, const string& chars) {
+string gen_string(RngT<Engine>& rng, Int n, const string& chars) {
     string res;
     res.reserve(n);
     REP(_, n) {
@@ -19,8 +19,8 @@ string gen_string(RngT<Engine>& rng, i64 n, const string& chars) {
 // n: 頂点数(n >= 2)
 // 辺のリストを返す
 template<typename Engine>
-vector<pair<i64,i64>> gen_tree(RngT<Engine>& rng, i64 n) {
-    vector<i64> prufer(n-2);
+vector<pair<Int,Int>> gen_tree(RngT<Engine>& rng, Int n) {
+    vector<Int> prufer(n-2);
     ALL(rng.generate, prufer, 0, n-1);
 
     return prufer_to_tree(prufer);
@@ -32,13 +32,13 @@ vector<pair<i64,i64>> gen_tree(RngT<Engine>& rng, i64 n) {
 // m: 辺数 (n-1 <= m <= n*(n-1)/2)
 // 辺のリストを返す
 template<typename Engine>
-vector<pair<i64,i64>> gen_connected_graph(RngT<Engine>& rng, i64 n, i64 m) {
+vector<pair<Int,Int>> gen_connected_graph(RngT<Engine>& rng, Int n, Int m) {
     ASSERT(n >= 2);
     ASSERT(n-1 <= m && m <= n*(n-1)/2);
 
-    set<pair<i64,i64>> es;
+    set<pair<Int,Int>> es;
     {
-        vector<pair<i64,i64>> ev = gen_tree(rng, n);
+        vector<pair<Int,Int>> ev = gen_tree(rng, n);
         for(auto& e : ev) {
             if(e.first > e.second)
                 swap(e.first, e.second);
@@ -47,8 +47,8 @@ vector<pair<i64,i64>> gen_connected_graph(RngT<Engine>& rng, i64 n, i64 m) {
     }
 
     REP(_, m-SIZE(es)) {
-        i64 s = -1;
-        i64 t = -1;
+        Int s = -1;
+        Int t = -1;
         while(s == t || es.find(make_pair(s,t)) != end(es)) {
             s = rng.uniform(0, n-1);
             t = rng.uniform(0, n-1);
@@ -57,7 +57,7 @@ vector<pair<i64,i64>> gen_connected_graph(RngT<Engine>& rng, i64 n, i64 m) {
         es.emplace(s,t);
     }
 
-    return vector<pair<i64,i64>>(begin(es), end(es));
+    return vector<pair<Int,Int>>(begin(es), end(es));
 }
 
 // ランダムな連結単純無向重み付きグラフを生成
@@ -68,17 +68,17 @@ vector<pair<i64,i64>> gen_connected_graph(RngT<Engine>& rng, i64 n, i64 m) {
 // cmax: 重み上限
 // 辺のリストを返す
 template<typename Engine>
-vector<tuple<i64,i64,i64>> gen_connected_graph_weighted(
-    RngT<Engine>& rng, i64 n, i64 m, i64 cmin, i64 cmax)
+vector<tuple<Int,Int,Int>> gen_connected_graph_weighted(
+    RngT<Engine>& rng, Int n, Int m, Int cmin, Int cmax)
 {
     ASSERT(cmin <= cmax);
 
     auto es = gen_connected_graph(rng, n, m);
 
-    vector<tuple<i64,i64,i64>> res;
+    vector<tuple<Int,Int,Int>> res;
     res.reserve(SIZE(es));
     for(const auto& e : es) {
-        i64 c = rng.uniform(cmin, cmax);
+        Int c = rng.uniform(cmin, cmax);
         res.emplace_back(e.first, e.second, c);
     }
