+++ include/prefix_function.hpp
@@ -1,12 +1,12 @@
 // prefix_function {{{
 
 template<class RandomIt>
-vector<i64> prefix_function(RandomIt first, RandomIt last) {
-    i64 n = last - first;
-    vector<i64> pi(n, 0);
+vector<Int> prefix_function(RandomIt first, RandomIt last) {
+    Int n = last - first;
+    vector<Int> pi(n, 0);
 
     FOR(i, 1, n) {
-        i64 j = pi[i-1];
+        Int j = pi[i-1];
         while(j > 0 && first[i] != first[j])
             j = pi[j-1];
         if(first[i] == first[j])
